requirements.txt
discord.py
aiohttp
python-dotenv
gTTS